fix(schema): prevent legacy data loss and make coercion server-safe
Follow-up fixes on top of the tenant-schema refactor:

- mergeSchemaDefaults: coerce the raw document BEFORE adding defaults.
  The previous order injected empty nested defaults (emailNotifications,
  attestations, resource.training) that made legacy documents look like
  the new nested shape, so coercion took its new-shape branch and let the
  empty defaults shadow real legacy values (emailMessages, agreements,
  needsSafetyTraining / training URLs) — silently blanking them.
- migrateResource: merge an existing nested `training` object with legacy
  flat fields so partially-migrated resources keep their form/info URLs.
- isNewSchemaShape: stop requiring the optional `contextLabels` field so a
  nested document that omits it is not misdetected as legacy and blanked.
- coerceTenantSchema: map legacy emailMessages and prefer non-empty
  attestations in the new-shape branch as well (defense in depth).
- FormInput: rename the inverted `agreementsChecked` to
  `attestationsIncomplete` to match what it actually represents.
- Extract server-safe schema types/defaults into `schemaTypes.ts` so API
  routes / coerceTenantSchema / firebase no longer pull the client-only
  SchemaProvider (createContext) into the server bundle. SchemaProvider
  now re-exports them and is marked "use client". Fixes the build.
- Add a regression test asserting legacy email/agreement/training values
  survive mergeSchemaDefaults.
