+++ booking-app/components/src/client/routes/components/SchemaProvider.tsx
@@ -1,408 +1,20 @@
-import { defaultSafetyTrainingInfoUrl } from "@/components/src/constants/safetyTraining";
+"use client";
+
 import { createContext, useContext } from "react";
+import {
+  generateDefaultSchema,
+  type SchemaContextType,
+} from "./schemaTypes";
 
-export { defaultSafetyTrainingInfoUrl };
+// Re-export all server-safe schema types/defaults so existing client consumers
+// can keep importing them from SchemaProvider. The actual definitions live in
+// `./schemaTypes` (no React), which lets server code import them without
+// dragging this client-only module into the server bundle.
+export * from "./schemaTypes";
 
-export type Attestation = {
-  id: string;
-  html: string;
-};
-
-/** @deprecated Use Attestation */
-export type Agreement = Attestation;
-
-export type StaffingSection = {
-  name: string;
-  indexes: number[];
-};
-
-export type ResourceTraining = {
-  required?: boolean;
-  formId?: string;
-  infoUrl?: string;
-};
-
-export type Resource = {
-  capacity: number;
-  name: string;
-  roomId: number;
-  isEquipment: boolean;
-  calendarId: string;
-  training?: ResourceTraining;
-  isWalkIn: boolean;
-  isWalkInCanBookTwo: boolean;
-  services: string[];
-  autoApproval?: {
-    shouldAutoApprove?: boolean;
-    minHour?: {
-      admin: number;
-      faculty: number;
-      student: number;
-    };
-    maxHour?: {
-      admin: number;
-      faculty: number;
-      student: number;
-    };
-    conditions?: {
-      setup: boolean;
-      equipment: boolean;
-      staffing: boolean;
-      catering: boolean;
-      cleaning: boolean;
-      security: boolean;
-    };
-  };
-  maxHour?: {
-    student: number;
-    faculty: number;
-    admin: number;
-    studentWalkIn: number;
-    facultyWalkIn: number;
-    adminWalkIn: number;
-    studentVIP: number;
-    facultyVIP: number;
-    adminVIP: number;
-  };
-  minHour?: {
-    student: number;
-    faculty: number;
-    admin: number;
-    studentWalkIn: number;
-    facultyWalkIn: number;
-    adminWalkIn: number;
-    studentVIP: number;
-    facultyVIP: number;
-    adminVIP: number;
-  };
-  staffingServices?: string[];
-  staffingSections?: StaffingSection[];
-  calendarIdProd?: string;
-};
-
-export type TimeSensitiveRequestWarning = {
-  hours?: number;
-  isActive?: boolean;
-  message?: string;
-  policyLink?: string;
-};
-
-export type ContextLabels = {
-  user: string;
-  worker: string;
-  reviewer: string;
-  services: string;
-  admin: string;
-};
-
-/** @deprecated Use ContextLabels */
-export type PermissionLabels = ContextLabels;
-
-export type TenantBranding = {
-  name: string;
-  logo: string;
-  nameForPolicy: string;
-  contextLabels: ContextLabels;
-};
-
-export type FormServicesConfig = {
-  showCatering: boolean;
-  showEquipment: boolean;
-  showSecurity: boolean;
-  showSetup: boolean;
-  showStaffing: boolean;
-};
-
-export type FormConfig = {
-  showBookingType: boolean;
-  showNNumber: boolean;
-  showSponsor: boolean;
-  services: FormServicesConfig;
-};
-
-export type OriginsConfig = {
-  VIP: boolean;
-  walkIn: boolean;
-};
-
-export type MappingsConfig = {
-  program: Record<string, string[]>;
-  role: Record<string, string[]>;
-  school: Record<string, string[]>;
-};
-
-export type TrainingConfig = {
-  formId?: string;
-};
-
-export type EmailNotifications = {
-  requestedUser: string;
-  requestedNeedsApproval: string;
-  reviewedNeedsApproval: string;
-  approvedWalkIn: string;
-  approvedVIP: string;
-  checkedOut: string;
-  checkedIn: string;
-  declined: string;
-  canceled: string;
-  canceledLate: string;
-  noShow: string;
-  closed: string;
-  approvedUser: string;
-};
-
-export type SchemaContextType = {
-  /** Tenant slug / document id (e.g. mc, itp) */
-  tenantId: string;
-  tenant: TenantBranding;
-  policy: string;
-  mappings: MappingsConfig;
-  roles: string[];
-  form: FormConfig;
-  attestations: Attestation[];
-  resources: Resource[];
-  origins: OriginsConfig;
-  training?: TrainingConfig;
-  supportPA?: boolean;
-  supportLiaison?: boolean;
-  resourceName: string;
-  declinedGracePeriod?: number;
-  /** Hours pending before interim column text is emphasized on the admin bookings dashboard (default 18) */
-  interimHighlightThresholdHours?: number;
-  /**
-   * When enabled, automatically cancel requests within a time window prior to start time
-   * if they are still unapproved (Requested / Pre-approved depending on conditions).
-   *
-   * Stored as `false` by default for backwards compatibility with older schemas.
-   */
-  autoCancel?:
-    | false
-    | {
-        minutesPriorToStart: number;
-        conditions: {
-          requested: boolean;
-          preApproved: boolean;
-        };
-      };
-  calendarConfig?: {
-    startHour?: Record<string, string>;
-    slotUnit?: Record<string, number>;
-    multipleResourceSelect?: boolean;
-    timeSensitiveRequestWarning?: TimeSensitiveRequestWarning;
-  };
-  ccEmails?: {
-    approved: { development: string; staging: string; production: string };
-    canceled: { development: string; staging: string; production: string };
-  };
-  emailNotifications: EmailNotifications;
-};
-
-function defineObjectArrayWithDefaults<T>(defaults: T): ObjectArrayWithDefaults<T> {
-  const value = [] as ObjectArrayWithDefaults<T>;
-  value.__defaults__ = defaults;
-  return value;
-}
-
-export interface ObjectArrayWithDefaults<T> extends Array<T> {
-  __defaults__: T;
-}
-
-export const defaultStaffingSection: StaffingSection = {
-  name: "",
-  indexes: [],
-};
-
-export const defaultAttestation: Attestation = {
-  id: "",
-  html: "",
-};
-
-/** @deprecated Use defaultAttestation */
-export const defaultAgreement = defaultAttestation;
-
-export const defaultResource: Resource = {
-  capacity: 0,
-  name: "",
-  roomId: 0,
-  isEquipment: false,
-  calendarId: "",
-  training: {
-    required: false,
-    formId: "",
-    infoUrl: defaultSafetyTrainingInfoUrl,
-  },
-  isWalkIn: false,
-  isWalkInCanBookTwo: false,
-  services: [],
-  autoApproval: {
-    shouldAutoApprove: false,
-    minHour: { admin: -1, faculty: -1, student: -1 },
-    maxHour: { admin: -1, faculty: -1, student: -1 },
-    conditions: {
-      setup: false,
-      equipment: false,
-      staffing: false,
-      catering: false,
-      cleaning: false,
-      security: false,
-    },
-  },
-  maxHour: {
-    student: -1,
-    faculty: -1,
-    admin: -1,
-    studentWalkIn: -1,
-    facultyWalkIn: -1,
-    adminWalkIn: -1,
-    studentVIP: -1,
-    facultyVIP: -1,
-    adminVIP: -1,
-  },
-  minHour: {
-    student: -1,
-    faculty: -1,
-    admin: -1,
-    studentWalkIn: -1,
-    facultyWalkIn: -1,
-    adminWalkIn: -1,
-    studentVIP: -1,
-    facultyVIP: -1,
-    adminVIP: -1,
-  },
-  staffingServices: [],
-  staffingSections: defineObjectArrayWithDefaults(defaultStaffingSection),
-  calendarIdProd: "",
-};
-
-const defaultTimeSensitiveRequestWarning: TimeSensitiveRequestWarning = {
-  hours: 48,
-  isActive: false,
-  message: "",
-  policyLink: "",
-};
-
-const defaultContextLabelsByTenantId = (tenantId?: string): ContextLabels => {
-  const normalized = (tenantId || "").toLowerCase();
-  if (normalized === "itp") {
-    return {
-      user: "User",
-      worker: "ER",
-      reviewer: "1st Approver",
-      services: "Services",
-      admin: "Admin",
-    };
-  }
-  return {
-    user: "User",
-    worker: "PA",
-    reviewer: "Liaison",
-    services: "Services",
-    admin: "Admin",
-  };
-};
-
-const defaultTenantBranding = (tenantId?: string): TenantBranding => ({
-  name: "",
-  logo: "",
-  nameForPolicy: "",
-  contextLabels: defaultContextLabelsByTenantId(tenantId),
-});
-
-export const defaultScheme: Omit<SchemaContextType, "tenantId"> = {
-  tenant: defaultTenantBranding(),
-  policy: "",
-  mappings: {
-    program: {},
-    role: {},
-    school: {},
-  },
-  roles: [],
-  form: {
-    showBookingType: true,
-    showNNumber: true,
-    showSponsor: true,
-    services: {
-      showCatering: true,
-      showEquipment: true,
-      showSecurity: true,
-      showSetup: true,
-      showStaffing: true,
-    },
-  },
-  attestations: defineObjectArrayWithDefaults(defaultAttestation),
-  resources: defineObjectArrayWithDefaults(defaultResource),
-  origins: {
-    VIP: false,
-    walkIn: false,
-  },
-  training: {
-    formId: "",
-  },
-  supportPA: false,
-  supportLiaison: false,
-  resourceName: "",
-  declinedGracePeriod: 24,
-  interimHighlightThresholdHours: 18,
-  autoCancel: false,
-  calendarConfig: {
-    startHour: {
-      student: "09:00:00",
-      studentVIP: "06:00:00",
-      studentWalkIn: "09:00:00",
-      faculty: "09:00:00",
-      facultyVIP: "06:00:00",
-      facultyWalkIn: "09:00:00",
-      admin: "09:00:00",
-      adminVIP: "06:00:00",
-      adminWalkIn: "09:00:00",
-    },
-    slotUnit: {
-      student: 15,
-      studentVIP: 15,
-      studentWalkIn: 15,
-      faculty: 15,
-      facultyVIP: 15,
-      facultyWalkIn: 15,
-      admin: 15,
-      adminVIP: 15,
-      adminWalkIn: 15,
-    },
-    multipleResourceSelect: false,
-    timeSensitiveRequestWarning: defaultTimeSensitiveRequestWarning,
-  },
-  ccEmails: {
-    approved: { development: "", staging: "", production: "" },
-    canceled: { development: "", staging: "", production: "" },
-  },
-  emailNotifications: {
-    requestedUser: "",
-    requestedNeedsApproval: "",
-    reviewedNeedsApproval: "",
-    approvedWalkIn: "",
-    approvedVIP: "",
-    checkedOut: "",
-    checkedIn: "",
-    declined: "",
-    canceled: "",
-    canceledLate: "",
-    noShow: "",
-    closed: "",
-    approvedUser: "",
-  },
-};
-
-export function generateDefaultSchema(tenantId: string): SchemaContextType {
-  return {
-    tenantId,
-    ...defaultScheme,
-    tenant: {
-      ...defaultScheme.tenant,
-      contextLabels: defaultContextLabelsByTenantId(tenantId),
-    },
-  };
-}
-
-export const SchemaContext = createContext<SchemaContextType>(generateDefaultSchema(""));
+export const SchemaContext = createContext<SchemaContextType>(
+  generateDefaultSchema(""),
+);
 
 export const useTenantSchema = () => useContext(SchemaContext);
